implement noise and blur source filters, move LPW in with other pipelines
